+++ src/nti/property/tests/test_tunables.py
@@ -3,14 +3,19 @@
 Tests for tunables.py
 
 """
-
 import unittest
 import doctest
 
-
-def test_suite():
-    # zope-testrunner specific test loading.
+def _doctest():
+    from zope.testing import cleanup
     from .. import tunables
-    return unittest.TestSuite((
-        doctest.DocTestSuite(tunables)
-    ))
+    return doctest.DocTestSuite(tunables,
+                                tearDown=lambda _test: cleanup.cleanUp())
+
+def load_tests(_loader, standard_tests, _pattern):
+    # unittest module
+    standard_tests.addTests(_doctest())
+    return standard_tests
+
+if __name__ == '__main__':
+    unittest.main()